Moved AlertDestinationsSection to new RelatedDestinations subfolder
