+++ web/src/components/cards/AlertCard/AlertCard.tsx
@@ -23,11 +23,11 @@ import SeverityBadge from 'Components/badges/SeverityBadge';
 import React from 'react';
 import urls from 'Source/urls';
 import LinkButton from 'Components/buttons/LinkButton';
+import AlertDestinationsSection from 'Components/RelatedDestinations/AlertDestinationsSection';
 import { AlertSummaryFull } from 'Source/graphql/fragments/AlertSummaryFull.generated';
 import { formatDatetime } from 'Helpers/utils';
 import BulletedLogType from 'Components/BulletedLogType';
 import UpdateAlertDropdown from '../../dropdowns/UpdateAlertDropdown';
-import AlertDestinationsSection from './AlertDestinationsSection';
 
 interface AlertCardProps {
   alert: AlertSummaryFull;
